moved token dropdown into separate component
